requirements.txt
Flask==3.0.0
requests==2.31.0
transformers==4.31.0
torch==2.2.0
google-generativeai==0.3.1
PyMuPDF==1.23.21
